Fix tests so that they can be run multiple times
One of the spec tests was creating a table to test
some user privileges.  However, it wasn't dropping
the table afterwards, meaning that the same test
would fail on the next run because that table already
existed.  This commit adds a command to drop the
table so that the tests can be run several times
in sequence.