register the new cluster after initdb to distribution scripts
Make the new created cluster know to Debian/RHEL scripts.

This will ensure that postgres will be restarted correctly on boot.

On Debian/Ubuntu if you are using postgresql-common => 142 it will also
allow to create the cluster in a non standard location disabling the
autocreation of cluster on install when run_initdb parameter is set