Change validate_postgresql_connection.sh exec PATH
FreeBSD needs /usr/local/bin in PATH in order to find bash. This does
require that the node has the bash port installed. It might be desired
that a separate script was provided for FreeBSD which used its /bin/sh,
this could be done by changing the for loop in the script to use

    for c in $(jot $TRIES)

in place of

    for (( c=1; c<=$TRIES; c++ ))