Changes to support OpenBSD.
Added $service_reload parameter to params.pp, in order to allow
a reload of the postgresql server on OpenBSD, which apparently doesn't
have a service binary.

Add $service_reload to the global variables.

document service_reload parameter in README.md

Add the service_reload parameter to the freshly added Suse OS.

undef is a keyword, not a string.
remove the firewall_supported parameter from OpenBSD too, as
it got removed from original master.

pick function doesn't work with undef, so don't use it,
and fix the default package name for the perl package.
Both just for OpenBSD.