Update server.pp
In @8db4a0c, the deprecation warning code was extended. This lead to deprecation warnings when you did _NOT_ specify the version in class { 'postgresql::server': }, because the parameter was always set by the inheritance down from postgresql::params::version (which is in turn set by postgresql::globals).

So in postgresql::server, this value could never ever be version==undef, so this if-statement was always true, leading to the "false positive". This wasn't catched by the unit tests.

Maybe I'll extend the unit test further, if I find the time...