Fix the handling of run_unless_sql_command in puppet 4
`Puppet::Util::SUIDManager.run_and_capture` (in puppet < 3.4) returns an
array of the output and the `Process::Status` object.
`Puppet::Util::Execution.execute` (in puppet >= 3.4) returns the output
and saves the `Process::Status` object to `$CHILD_STATUS`

This is handled in the provider, but the type was also trying to handle
it, and was not last updated when the provider was.