(MODULES-3319) Hash argument for create_resources #765
Without this patch, seperate modules and classes need to be built
to declare the provided defines for the postgresql server.
This patch permits the passing of a hash into the postgresql::server
function for the various resources.

The puppetlabs-mysql module provides similar behavior whereby
databases can be managed as values from Hiera or an ENC. This
commit brings the puppetlabs-postgresql module into feature
parity with the puppetlabs-mysql module.