(MODULES-1394) replace validate_db_connection type with custom type
This replaces the defined type `validate_db_connection` with a new
custom type called `postgresql_conn_validator`. It's functionality is
nearly a copy of its predecessor with the exception of the
`create_db_first` parameter which is taken care of with a collector in
service.pp. The old type is still intact but all docs have been removed
and a warning has been attached to it announcing its deprecation.