(maint) mock structured facts
we recently switched apt back to the params pattern and it now uses structured facts. since postgres includes apt, some postgres tests needed to have structured facts mocked out because most of the tests still use legacy facts. the mocking gets redundant and we should make an effor to go in at some point and switch postgres to structured facts completely but this works for now.