Add RDS conditional
For our migration to AWS we're managing users and databases using this
module but using an RDS instance.

Unfortunately, due to the permissions we're unable to do things because
"pg_shadow" doesn't exist.

See this (currently open) ticket for reference: https://tickets.puppetlabs.com/browse/MODULES-5068

We're a couple of versions behind in the module, so for the time-being
we're going to have to fork it and add a conditional so we don't affect
the running of our current servers in our current hosting provider.