(MODULES-3804) Fix sort order of pg_hba_rule entries
The `$order`-parameter to `pg_hba_rule` can be either an integer or a
string, but the corresponding `concat`-resource is configured to sort
the entries using alphanumeric sort.

The result is that a `pg_hba_rule` entry with `order => 2` sorts after
an entry with `order => 10`.

This is also a problem with the default `pg_hba_rule` entries, which
uses numeric order values 1, 2, 3, 4, 100 and 101. These are sorted in
the incorrect order (1, 100, 101, 2, 3, 4).

Unfortunately, since string values are allowed in the
`$order`-parameter, we cannot simply change the `concat`-resource to
use numeric sort.

This patch fixes this issue by zero-padding all integer values to
three digits. E.g. 1 becomes `001`, and 10 becomes `010`. This should
work well for integer values from 0 through 999, and works well with
earlier versions of the module which used string values in that range
for the `$order`-parameter.