Finish conversion of `postgresql_password` function
In https://github.com/puppetlabs/puppetlabs-postgresql/pull/1129 the 3
functions from this module were ported to the modern function API and
namespaced.  None of the boilerplate was fixed up, tests weren't
updated, documentation wasn't updated etc.

This commit finishes the work for the `postgresql_password` function.

* Remove old API `puppet/parser/functions` version.
* Create a replacement non-namespaced, (but deprecated), new API shim
that calls the namespaced version.
* Replace argument validation in the function code with data-type based
validation for the individual function parameters.
* Update the README and REFERENCE.md files.
* Test both shim and namespaced version of the function.