Manage postgresql_conf_path file permissions
This adds a parameter (defaulting to true) to manage the
postgresql_conf_path file resource. The actual content isn't managed
because that's what postgresql::server::config_entry is for. However,
managing the file ensures the correct permissions.

It also provides a workaround for PUP-10548[1] when using Red Hat SCL
for packages when the correct SELinux file context isn't present.
Without managing the file, an admin will need to either make sure the
package is present before running Puppet, manage the file via another
module or manually set the file context after Puppet ran. With this
workaround, Puppet will converge on the second run and actually start
PostgreSQL.

The acceptance test path is changed to avoid problems on Debian(-based)
systems where the root user can't write to it. Root doens't own the file
and doesn't have write access due to the sticky bit.

[1]: https://tickets.puppetlabs.com/browse/PUP-10548