Change - Use systemd drop-in directory for override settings
The .include directive has been deprecated and support for it may
be removed in a future systemd release.  Drop-in directories are
supported on all current RedHat, Fedora, and Gentoo releases.

This branch contains changes to support this feature and disables
the old behavior of using Include directives on RedHat based nodes.