(maint) Fix warning introduced in 08a1752
`unless` does not accept a Sensitive parameter, and passing one result
in a warning being output:

```
(/Postgresql_psql[ALTER ROLE puppetdb ENCRYPTED PASSWORD ****]) Unable to mark 'unless' as sensitive: unless is a parameter and not a property, and cannot be automatically redacted.
```

This is not really an issue here since this query is not displayed in
the logs, even in debug mode.