Remove use of spec_helper_acceptance in unit tests
In c96b073b40b7d48ac5c7a4e68796b4f87ba65abd the acceptance helper was
included in the unit tests. This should never happen and breaks the
tests. It appears it was included to have a postgresql_version but that
should be defined in let blocks. The acceptance helper only works if
there is an actual target machine.

Fixes: c96b073b40b7d48ac5c7a4e68796b4f87ba65abd