Inline file contents in the catalog
The file() function can read a file form a module and return the
content. This happens server side and is included in the catalog. This
saves a request to the fileserver at runtime. That request can fail for
various reasons (it's still a network). It also means that a cached
catalog is sufficient and an agent can reapply it without a Puppetserver
running. These files are small enough to inline in the catalog.