Add Red Hat Enterprise Linux 9 support
The comment indicates the checks in server config need to be done on EL6
and older only. The current check would also trigger on EL9. This uses
a version comparison on the major version. We can leave out the Fedora
check since we don't need to care about Fedora < 7.

It also sets the default version on EL9 to version 13. This should also
affect CentOS Stream 9, which is supposed to work but not officially
verified to work.