Add multi instance support, refactoring late_initdb.pp
This commit adds changes a class to add multi instance support to this
module. The general idea is to first copy all classes which are used and
create defines from them. These classes will use the defines as is.
Necessary changes for the instances itself will be added to the classes
and defined types at a later point. This ensures, the module will work
as it does right now and there are no breaking changes.