Fix `postgresql::default()` return value for unknown parameters
When calling `postgresql::default()` with a name that has no corresponding
parameter, the function returns the string
"postgresql::globals::<parameter-name>" which is probably not intended
given the comment above the code.

The usage of pick seems to indicate that an exception should be raised
when a parameter is not found in `params` nor `globals.pp`, so adjust
the code accordingly.