Don't use type alias in `postgresql_password`
Type aliases from modules aren't available on the agent.
See https://puppet.atlassian.net/browse/PUP-7197

This means that calls to `postgresql::postgresql_password` can't be
deferred if the `hash` parameter uses the type
`Postgresql::Pg_password_encryption`. Instead we have to use the raw
`Enum`, (which unfortunately causes code duplication, but this is
unavoidable).

Using this function with `Deferred` is not just a theoretical use-case.
The module itself tries to here.
https://github.com/puppetlabs/puppetlabs-postgresql/blob/20704ffae24dcf970784697b1798c09d026fb7f8/manifests/server/role.pp#L162