Debian 11: Use correct package name for python bindings
Since Debian 11 was added to this module, the package name was wrong. It
was set to python-psycopg2, but the correct name is python3-psycopg2.

There's a longer analysis about this bug in https://github.com/puppetlabs/puppetlabs-postgresql/pull/1638/files#r2390588710

tl;dr:

* Debian 10 used python-psycopg2 and that was set by the module
* Debian 11 was added without proper acceptance testing
* Later on unit tests were added for Debian 11, but still with the wrong
  package name
* 651c2b4ec0a3d9a76770369625b0a91bcba444ff fixed the package name, but
  in the wrong place and didn't adjust the tests
* That resulted in a broken pipeline, that's how I found all of this

This patch:
* Cleans up the hiera data
* sets the package name at the correct location
* adjusts the debian 11 unit test for the new package name
* Adds tests for Debian 12 and 13 (won't be executed right now because
  Perforce ships a too old version of FacterDB)

Edit: Had to remove Debian 13 because the pipeline fails if it cannot
find a factset.